Encapsulate User Tools JavaScript into a Stimulus controller, replacing any jQuery
