+++ custom/styles/_custom.scss
@@ -9,6 +9,10 @@
   z-index: 1;
 }
 
+#me .fa-spinner.hide {
+  display: none;
+}
+
 // =General use
 
 .wrapper {
